DL-4363: Show error to user when enrolment call fails
