+++ app/services/SubscriptionService.scala
@@ -43,20 +43,17 @@ class SubscriptionService @Inject()(metrics: AwrsMetrics,
     val timer = metrics.startTimer(ApiType.API4Subscribe)
     for {
       submitResponse <- etmpConnector.subscribe(data, safeId)
-      ggResponse <- addKnownFacts(submitResponse, safeId, utr, businessType, postcode)
+      enrolmentResponse <- addKnownFacts(submitResponse, safeId, utr, businessType, postcode)
     } yield {
-      ggResponse.status match {
-        case OK =>
+      (submitResponse.status, enrolmentResponse.status) match {
+        case (OK, NO_CONTENT) =>
           timer.stop()
           metrics.incrementSuccessCounter(ApiType.API4AddKnownFacts)
           submitResponse
         case _ =>
           timer.stop()
           metrics.incrementFailedCounter(ApiType.API4AddKnownFacts)
-          // Code changed to always return the etmp response even if there is a GG failure.
-          // The GG failure will need to be sorted out manually and there is nothing the user can do at the time.
-          // The manual process will take place after the GG failure is picked up in Splunk.
-          submitResponse
+          enrolmentResponse
       }
     }
   }
@@ -77,7 +74,7 @@ class SubscriptionService @Inject()(metrics: AwrsMetrics,
         val enrolmentKey = s"$AWRS_SERVICE_NAME~AWRSRefNumber~$awrsRegistrationNumber"
         val enrolmentVerifiers = createVerifiers(safeId, utr, businessType, postcode)
         enrolmentStoreConnector.upsertEnrolment(enrolmentKey, enrolmentVerifiers)
-      case _ => Future.successful(response)
+      case _ => Future(response)
     }
 
   private def createVerifiers(safeId: String, utr: Option[String], businessType: String, postcode: String) = {
