+++ it/uk/gov/hmrc/helpers/http/StubbedBasicHttpCalls.scala
@@ -6,28 +6,8 @@ import com.github.tomakehurst.wiremock.stubbing.StubMapping
 
 trait StubbedBasicHttpCalls {
 
-  def stubbedHead(url: String, statusCode: Int): StubMapping = {
-    stubFor(head(urlMatching(url))
-      .willReturn(
-        aResponse()
-          .withStatus(statusCode)
-      )
-    )
-  }
-
   def stubbedGet(url: String, statusCode: Int, responseBody: String): StubMapping = {
     stubFor(get(urlPathMatching(url))
-      .willReturn(
-        aResponse()
-          .withStatus(statusCode)
-          .withBody(responseBody)
-      )
-    )
-  }
-
-  def  stubbedGetQueryParams(url: String, statusCode: Int, responseBody: String): StubMapping = {
-    stubFor(get(urlPathEqualTo(url))
-      .withQueryParam("encoding-type", equalTo("url"))
       .willReturn(
         aResponse()
           .withStatus(statusCode)
@@ -48,25 +28,6 @@ trait StubbedBasicHttpCalls {
 
   def stubbedPut(url: String, statusCode: Int): StubMapping = {
     stubFor(put(urlMatching(url))
-      .willReturn(
-        aResponse()
-          .withStatus(statusCode)
-      )
-    )
-  }
-
-  def stubbedPatch(url: String, statusCode: Int, responseBody: String = ""): StubMapping = {
-    stubFor(patch(urlMatching(url))
-      .willReturn(
-        aResponse()
-          .withStatus(statusCode)
-          .withBody(responseBody)
-      )
-    )
-  }
-
-  def stubbedDelete(url: String, statusCode: Int): StubMapping = {
-    stubFor(delete(urlMatching(url))
       .willReturn(
         aResponse()
           .withStatus(statusCode)
